fix mouse does not work when minimized

+++ workspace/Sep_10_Laser/src/MouseRobot.java
@@ -1,3 +1,8 @@
 public interface MouseRobot {
-    void click(int x, int y);
+    /**
+     * Click mouse.
+     * @param x percentage of screen width, 0-1
+     * @param y percentage of screen height, 0-1
+     */
+    void click(double x, double y);
 }
